dotgit: fix a filesystem race in Refs/walkReferencesTree
In walkReferencesTree(), the filesystem is browsed recursively. After a folder
is listed, each child element (directory, file) are inspected.

What can happen is that by the time we get to operate on the child element, it
might have been deleted from the filesystem and we would get a PathError.

In the case of directories there was already a case to avoid bubbling up the error
(we consider that there is no ref there, which is correct), but that was missing
for files. This commit just apply the same logic.